Split cvat_sdk/pytorch/__init__.py into multiple files
I'm starting to get lost in it.
